+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
+## 3.0.0
+  * Fix dynamic stream discovery and sync [#38](https://github.com/singer-io/tap-quickbase/pull/38)
+  
 ## 2.1.1
   * Improved 5xx error handling [37](https://github.com/singer-io/tap-quickbase/pull/37)
 
 ## 2.1.0
-  * Complete rewrite using streams-based architecture [#29](https://github.com/singer-io/tap-quickbase/pull/29)
+  * Add static stream discovery and sync [#29](https://github.com/singer-io/tap-quickbase/pull/29)
   * Upgrade to Python 3.12
   * Update dependencies: singer-python 6.3.0, backoff 2.2.1
   * Add unit and integration tests